requirements.txt
openai
python-dotenv
langchain
client
elevenlabs